Created otel java lib and publishing conventions
