+++ buildSrc/src/main/kotlin/otel.java-library-conventions.gradle.kts
@@ -1,0 +1,4 @@
+plugins {
+    id("java-library")
+    id("otel.errorprone-conventions")
+}
